vertical-demos/hospital-visit-summary/requirements.txt
jpype1
jaydebeapi
pandas
streamlit
langchain
langchain_nvidia_ai_endpoints
python-dotenv
